Move PaletteItemDescriptor to UI project
